pyproject.toml
[project]
name = "wikon2mqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.1.0",
    "schedule>=1.2.2",
    "selenium>=4.31.0",
]
